git_file_update/requirements.txt
certifi==2019.6.16
chardet==3.0.4
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
requests==2.22.0
smmap2==2.0.5
urllib3==1.25.3